Parsing math(v1.0)
Getting command line argument in a form od math expression, later on I will be forming postfix based on that. For now parse the input using sstream, checking if a token (value) is a number or no (! problem with expressions like 1+2, as it is seen as one token. For now separates tokens: numbers to queue, operators to stack. later on: 1) add precedence; 2) based on that form resulting postfix queue; 3) capture all this fuctionality in Postfix class