Full postfix with unary and power
In ToPostfix we first go and parse the input, meaning selete spaces, search for unary minus and replace it with char u. That's it. Then we go and look if any part of the string cannot go further solo, if it is a number like 11, or -1. Added ^ for raised to power, and it has one more if when meeting with the stack.